+++ tests/testthat/test-diffnet-class.R
@@ -1,6 +1,64 @@
 context("Data-management on diffnet objects")
 
-test_that("Subsetting slices", {
+# Checking attributes ----------------------------------------------------------
+test_that("Checking attributes in as_diffnet", {
+  # Generating data
+  set.seed(3312)
+  graph <- lapply(1990:2001, function(x) rgraph_ba(t = 9))
+  attrs <- matrix(runif(10*3), ncol=3)
+  dynat <- lapply(1990:2001, function(x) attrs)
+  toa   <- sample(c(NA,1990:2001), 10, TRUE)
+
+  # Dynamic attributes --------------------------------------------------------
+  # Number of rows
+  expect_error(
+    as_diffnet(graph, toa, t0=1990, t1=2001, vertex.dyn.attrs = attrs[-1,]),
+    "incorrect number of rows")
+
+  # Length of list
+  expect_error(
+    as_diffnet(graph, toa, t0=1990, t1=2001, vertex.dyn.attrs = dynat[-1]),
+    "equal.+slices")
+
+  # An element has different length
+  expect_error({
+    dynat[[2]] <- dynat[[1]][-1,]
+    dynat[[4]] <- dynat[[1]][-1,]
+    as_diffnet(graph, toa, t0=1990, t1=2001, vertex.dyn.attrs = dynat)
+  }, "don't have n.+rows")
+  dynat <- lapply(1990:2001, function(x) attrs)
+
+  # Elements have different classes
+  expect_error({
+    dynat[[6]] <- as.data.frame(dynat[[6]])
+    as_diffnet(graph, toa, t0=1990, t1=2001, vertex.dyn.attrs = dynat)
+  }, "should be of the same class")
+  dynat[[6]] <- dynat[[1]]
+
+  # Have different colnames
+  cnames <- netdiffuseR:::make_col_names(3, TRUE)
+  dynat <- lapply(dynat, function(x) {
+    colnames(x) <- cnames
+    x
+  })
+  colnames(dynat[[4]]) <- paste0(colnames(dynat[[1]]),"col")
+  expect_error({
+    as_diffnet(graph, toa, t0=1990, t1=2001, vertex.dyn.attrs = dynat)
+  }, "have the same colname")
+
+
+  # Dynamic list of vectors
+  dynat <- lapply(1990:2001, function(x) attrs[,1])
+  as_diffnet(graph, toa, t0=1990, t1=2001, vertex.dyn.attrs = dynat)
+
+  # Static attributes ----------------------------------------------------------
+  expect_error(
+    as_diffnet(graph, toa, t0=1990, t1=2001, vertex.static.attrs = attrs[-1,]),
+    "incorrect number of rows")
+})
+
+# Summary ----------------------------------------------------------------------
+test_that("Summary and subsetting slices", {
   set.seed(13131)
   diffnet <- rdiffnet(100, 20)
 
@@ -11,6 +69,7 @@ test_that("Subsetting slices", {
   expect_equal(sum1[nrow(sum1),c("cum_adopt")],sum2[nrow(sum2),c("cum_adopt")])
 })
 
+# diffnet.attrs errors ---------------------------------------------------------
 test_that("Error messages", {
   set.seed(009)
   diffnet <- rdiffnet(80, 20, seed.nodes = "random", seed.p.adopt = .1)